requirements.txt
# Web framework
fastapi==0.135.1
uvicorn[standard]
python-multipart

# Data & ML
pandas==3.0.1
numpy==2.4.3
scikit-learn==1.8.0
xgboost==3.2.0
joblib==1.5.3
scipy
statsmodels
feature_engine
openpyxl

# LLM providers
groq==1.1.1
google-genai

# Pydantic & config
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.2

# HTTP
httpx
requests